refactor: split types/rezervo into separate files
