+++ src/components/modals/ClassInfo/ClassInfoModal.tsx
@@ -3,8 +3,9 @@ import React, { Dispatch, SetStateAction } from "react";
 
 import ClassInfo from "@/components/modals/ClassInfo/ClassInfo";
 import { classRecurrentId } from "@/lib/helpers/recurrentId";
+import { AllConfigsIndex } from "@/types/config";
+import { RezervoClass } from "@/types/integration";
 import { ClassPopularity, ClassPopularityIndex } from "@/types/popularity";
-import { AllConfigsIndex, RezervoClass } from "@/types/rezervo";
 
 const ClassInfoModal = ({
     classInfoClass,
