+++ src/components/schedule/WeekSchedule.tsx
@@ -3,8 +3,9 @@ import { useRouter } from "next/router";
 import React, { useEffect } from "react";
 
 import DaySchedule from "@/components/schedule/DaySchedule";
+import { AllConfigsIndex } from "@/types/config";
+import { RezervoClass, RezervoWeekSchedule } from "@/types/integration";
 import { ClassPopularityIndex } from "@/types/popularity";
-import { AllConfigsIndex, RezervoClass, RezervoWeekSchedule } from "@/types/rezervo";
 
 function WeekSchedule({
     weekSchedule,
