+++ src/components/schedule/class/ClassCard.tsx
@@ -11,8 +11,10 @@ import { useUserSessions } from "@/lib/hooks/useUserSessions";
 import { randomElementFromArray } from "@/lib/utils/arrayUtils";
 import { hexColorHash, hexWithOpacityToRgb } from "@/lib/utils/colorUtils";
 import { EnterLeaveAnimation, OVER_THE_TOP_ANIMATIONS } from "@/types/animation";
+import { UserNameWithIsSelf } from "@/types/config";
+import { RezervoClass } from "@/types/integration";
 import { ClassPopularity } from "@/types/popularity";
-import { RezervoClass, SessionStatus, StatusColors, UserNameWithIsSelf } from "@/types/rezervo";
+import { SessionStatus, StatusColors } from "@/types/userSessions";
 
 const ClassCard = ({
     _class,
