+++ src/lib/helpers/popularity.ts
@@ -1,7 +1,7 @@
 import { isClassInThePast } from "@/lib/helpers/date";
 import { classRecurrentId } from "@/lib/helpers/recurrentId";
+import { RezervoClass, RezervoWeekSchedule } from "@/types/integration";
 import { ClassPopularity, ClassPopularityIndex } from "@/types/popularity";
-import { RezervoClass, RezervoWeekSchedule } from "@/types/rezervo";
 
 export function determineClassPopularity(_class: RezervoClass) {
     if (!_class || _class.availableSlots === undefined) return ClassPopularity.Unknown;
