+++ src/lib/hooks/useIntegrationUser.ts
@@ -5,7 +5,7 @@ import useSWRMutation from "swr/mutation";
 import { useUserConfig } from "@/lib/hooks/useUserConfig";
 import { IntegrationIdentifier } from "@/lib/integrations/active";
 import { fetcher } from "@/lib/utils/fetchUtils";
-import { IntegrationUser, IntegrationUserPayload } from "@/types/rezervo";
+import { IntegrationUser, IntegrationUserPayload } from "@/types/config";
 
 function putIntegrationUser(url: string, { arg: integrationUser }: { arg: IntegrationUserPayload }) {
     return fetch(url, {
