+++ src/types/integration.ts
@@ -2,75 +2,6 @@ import { DateTime } from "luxon";
 
 import { IntegrationIdentifier } from "@/lib/integrations/active";
 
-export type ClassTimeConfig = {
-    hour: number;
-    minute: number;
-};
-
-export type ClassConfig = {
-    activity: number;
-    display_name: string;
-    weekday: number;
-    studio: number;
-    time: ClassTimeConfig;
-};
-
-export type IntegrationConfig = {
-    active: boolean;
-    classes: ClassConfig[];
-};
-export type IntegrationConfigPayload = IntegrationConfig;
-
-export type IntegrationUser = {
-    username: string;
-};
-
-export type IntegrationUserPayload = IntegrationUser & {
-    password: string;
-};
-
-export type NotificationsConfig = {
-    reminder_hours_before: number | null;
-};
-
-export type Preferences = {
-    notifications: NotificationsConfig | null;
-};
-
-export type PreferencesPayload = Preferences;
-
-export type UserNameWithIsSelf = {
-    is_self: boolean;
-    user_name: string;
-};
-
-export type AllConfigsIndex = {
-    [recurrentClassId: string]: UserNameWithIsSelf[];
-};
-
-export enum SessionStatus {
-    BOOKED = "BOOKED",
-    CONFIRMED = "CONFIRMED",
-    WAITLIST = "WAITLIST",
-    UNKNOWN = "UNKNOWN",
-}
-
-export enum StatusColors {
-    ACTIVE = "#44b700",
-    WAITLIST = "#b75f00",
-    UNKNOWN = "#000",
-}
-
-export type UserNameSessionStatus = {
-    is_self: boolean;
-    user_name: string;
-    status: SessionStatus;
-};
-
-export type UserSessionsIndex = {
-    [classId: string]: UserNameSessionStatus[];
-};
-
 export type IntegrationProfile = {
     acronym: IntegrationIdentifier;
     name: string;
@@ -129,10 +60,4 @@ export type RezervoActivity = {
 
 export type IntegrationPageParams = {
     integration: IntegrationIdentifier;
-};
-
-export type RezervoCategory = {
-    name: string;
-    color: string;
-    keywords: string[];
 };
